Successfully refactored functions to sub-folders and migrated shared models to higher level file structure
