+++ functions/package.json
@@ -15,7 +15,7 @@
   "engines": {
     "node": "12"
   },
-  "main": "lib/index.js",
+  "main": "lib/functions/src/index.js",
   "dependencies": {
     "@firebase/app": "^0.6.10",
     "@google-cloud/storage": "^5.3.0",
