+++ shared_models/index.ts
@@ -1,0 +1,7 @@
+export interface IKeyMap<T> {
+  [key: string]: T;
+}
+
+export { ArticleDetailI, ArticlePreviewI } from './article.models';
+
+export { CommentI, ParentTypesE, VoteDirectionsE } from './comment.models';
